zconf.h: Prevent incorrect detection of unistd.h
A project might define HAVE_UNISTD_H to 0 and then `#include <zlib.h>`
which will lead to zconf.h incorrectly assuming that unistd.h exists
because HAVE_UNISTD_H is not undefined. F.ex., ffmpeg does this.

To keep the impact of this minimal since projects might rely on zlib.h
automatically including unistd.h this way, only do this on MSVC where
we know it is not available.